add tox.ini and use pytest
