+++ .travis.yml
@@ -1,0 +1,11 @@
+language: python
+sudo: false
+python: 2.7
+env:
+  matrix:
+    - TOXENV=py27-django18
+    - TOXENV=py27-django19
+
+install:
+  - pip install tox
+script: tox
