+++ README.rst
@@ -3,7 +3,7 @@ django-newauth
 
 See the docs at: https://github.com/beproud/django-newauth/tree/master/docs/en/source
 
-Run tests using setup.py::
+Run tests using tox::
 
-    $ python setup.py test
+    $ tox
