--- newauth/tests/__init__.py
@@ -1,6 +1,0 @@
-from backend_tests import *
-from api_tests import *
-from middleware_tests import *
-from decorator_tests import *
-from views_tests import *
-from testcase_tests import *
