--- requirements.txt
@@ -1,1 +1,0 @@
-Django>=1.1.1
