+++ tests/conftest.py
@@ -1,0 +1,7 @@
+import os
+
+os.environ['DJANGO_SETTINGS_MODULE'] = 'settings'
+
+
+from django.apps import apps
+apps.populate(['testapp'])
