+++ tests/templates/registration/login.html
@@ -1,0 +1,3 @@
+<form action="{% url 'newauth_login' %}" method="POST">
+{{ form.as_ul }}
+</form>
