+++ tox.ini
@@ -1,0 +1,18 @@
+# Requires tox > 1.8
+
+[tox]
+envlist = py27-django{18,19}
+
+[testenv]
+basepython = python2.7
+deps =
+    pytest
+    pytest-django
+    django18: Django>=1.8,<1.9
+    django19: Django>=1.9,<1.10
+    django110: Django>=1.10,<1.11
+setenv =
+    DJANGO_SETTINGS_MODULE = settings
+    PYTHONPATH = {toxinidir}/tests
+commands=py.test tests {posargs}
+
